backend/requirements.txt
# ============================================
# CORE WEB FRAMEWORK
# ============================================
fastapi[standard]==0.118.0
uvicorn[standard]==0.37.0
starlette==0.48.0
pydantic==2.10.6  # Versión compatible con pydantic-core
pydantic-settings==2.7.1
python-multipart==0.0.20
python-dotenv==1.1.1

# ============================================
# DATABASE & ORM
# ============================================
sqlalchemy==2.0.43
psycopg2-binary==2.9.10  # PostgreSQL adapter
alembic==1.16.5  # Database migrations
asyncpg==0.30.0  # Async PostgreSQL driver

# ============================================
# CACHE & REAL-TIME
# ============================================
redis==5.2.1  # Redis client
hiredis==3.2.1  # Requiere compilador C++ - Opcional para performance

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.5.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
bcrypt==5.0.0
PyJWT==2.10.1

# ============================================
# MACHINE LEARNING & IA
# ============================================
# Core ML libraries (versiones con wheels para Python 3.13)
numpy==2.3.3
pandas>=2.2.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0

# Deep Learning - PyTorch tiene wheels para Python 3.13
torch>=2.5.0,<3.0.0
torchvision>=0.20.0,<1.0.0

# ML utilities
joblib>=1.4.0,<2.0.0

# ============================================
# LLM & CHATBOT (Ollama via Docker)
# ============================================
aiohttp==3.11.11  # Async HTTP client para comunicación con Ollama
httpx==0.28.1  # Async HTTP client alternativo
# Ollama se ejecutará en Docker, no necesita cliente Python instalado

# ============================================
# WEBSOCKET & REAL-TIME
# ============================================
websockets==15.0.1
# python-socketio==5.12.0  # Socket.IO (opcional)

# ============================================
# VALIDATION & SERIALIZATION
# ============================================
email-validator==2.3.0
phonenumbers==9.0.15  # Validación de teléfonos
annotated-types==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0

# ============================================
# HTTP & ASYNC
# ============================================
httpcore==1.0.9
httptools==0.6.4
h11==0.16.0
anyio==4.11.0
sniffio==1.3.1

# ============================================
# TESTING
# ============================================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
faker==33.1.0  # Datos fake para testing
factory-boy==3.3.1  # Factories para testing

# ============================================
# UTILITIES
# ============================================
python-dateutil==2.9.0.post0  # Manejo de fechas
pytz==2024.2  # Timezones
# uuid==1.30  # Incluido en Python standard library, no necesario instalar
click==8.3.0  # CLI tools
colorama==0.4.6  # Colored terminal output
rich==14.1.0  # Beautiful terminal output
Pygments==2.19.2

# ============================================
# MONITORING & LOGGING
# ============================================
# sentry-sdk==2.39.0  # Error tracking
# loguru==0.7.3  # Advanced logging

# ============================================
# API DOCUMENTATION & TOOLS
# ============================================
Jinja2==3.1.6
MarkupSafe==3.0.3
markdown-it-py==4.0.0
mdurl==0.1.2
PyYAML==6.0.3

# ============================================
# DEVELOPMENT TOOLS
# ============================================
watchfiles==1.1.0  # Auto-reload
fastapi-cli==0.0.13
typer==0.19.2
shellingham==1.5.4
rignore==0.6.4

# ============================================
# NETWORKING & DNS
# ============================================
certifi==2025.8.3
urllib3==2.5.0
idna==3.10
dnspython==2.8.0

# ============================================
# OPTIONAL: API RATE LIMITING
# ============================================
# slowapi==0.1.9  # Rate limiting para FastAPI
